Rename pipeline.rs -> world.rs, world.rs -> manager.rs, MotionGfxWorld -> MotionGfxManager
