+++ crates/bevy_motiongfx/README.md
@@ -40,7 +40,7 @@ use bevy_motiongfx::prelude::*;
 
 fn build_timeline(
     mut commands: Commands,
-    mut motiongfx: ResMut<MotionGfxWorld>,
+    mut motiongfx: ResMut<MotionGfxManager>,
 ) {
     // Spawn the Entity.
     let entity = commands
@@ -72,7 +72,7 @@ use bevy_motiongfx::prelude::*;
 
 fn build_timeline(
     mut commands: Commands,
-    mut motiongfx: ResMut<MotionGfxWorld>,
+    mut motiongfx: ResMut<MotionGfxManager>,
     mut materials: ResMut<Assets<StandardMaterial>>
 ) {
     // Create the asset.
@@ -112,7 +112,7 @@ use bevy_motiongfx::prelude::*;
 
 fn build_timeline(
     mut commands: Commands,
-    mut motiongfx: ResMut<MotionGfxWorld>,
+    mut motiongfx: ResMut<MotionGfxManager>,
 ) {
     // Build the timeline.
     let mut b = motiongfx.create_bevy_builder();
