+++ crates/bevy_motiongfx/src/controller.rs
@@ -3,7 +3,7 @@ use bevy_ecs::prelude::*;
 use bevy_time::prelude::*;
 
 use crate::MotionGfxSet;
-use crate::world::{MotionGfxWorld, TimelineId};
+use crate::manager::{MotionGfxManager, TimelineId};
 
 pub struct ControllerPlugin;
 
@@ -18,7 +18,7 @@ impl Plugin for ControllerPlugin {
 }
 
 fn realtime_player_timing(
-    mut motiongfx: ResMut<MotionGfxWorld>,
+    mut motiongfx: ResMut<MotionGfxManager>,
     q_timelines: Query<(&TimelineId, &RealtimePlayer)>,
     time: Res<Time>,
 ) {
@@ -35,7 +35,7 @@ fn realtime_player_timing(
 }
 
 fn fixed_rate_player_timing(
-    mut motiongfx: ResMut<MotionGfxWorld>,
+    mut motiongfx: ResMut<MotionGfxManager>,
     mut q_timelines: Query<(&TimelineId, &mut FixedRatePlayer)>,
 ) {
     for (id, mut player) in
