+++ crates/bevy_motiongfx/src/world.rs
@@ -1,212 +1,65 @@
-use core::ops::{Deref, DerefMut};
-
-use bevy_app::prelude::*;
+use bevy_ecs::component::Mutable;
 use bevy_ecs::prelude::*;
-use bevy_platform::collections::HashMap;
 use motiongfx::prelude::*;
 
-use crate::MotionGfxSet;
-use crate::controller::FixedRatePlayer;
-use crate::controller::RealtimePlayer;
-use crate::pipeline::{BevyTimeline, BevyWorld};
-use crate::prelude::BevyTimelineBuilder;
+/// Newtype wrapper around [`World`] that is local to this crate,
+/// allowing [`SubjectSource`] impls without violating the orphan rule.
+#[repr(transparent)]
+pub struct BevyWorld(pub World);
 
-pub struct MotionGfxWorldPlugin;
+impl BevyWorld {
+    pub fn from_ref(world: &World) -> &Self {
+        // SAFETY: `BevyWorld` is repr(transparent) over `World`.
+        unsafe { &*(world as *const World as *const Self) }
+    }
 
-impl Plugin for MotionGfxWorldPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<MotionGfxWorld>().add_systems(
-            PostUpdate,
-            (
-                sample_timelines.in_set(MotionGfxSet::Sample),
-                (
-                    complete_timelines::<RealtimePlayer>,
-                    complete_timelines::<FixedRatePlayer>,
-                )
-                    .after(MotionGfxSet::Sample),
-            ),
-        );
+    pub fn from_mut(world: &mut World) -> &mut Self {
+        // SAFETY: `BevyWorld` is repr(transparent) over `World`.
+        unsafe { &mut *(world as *mut World as *mut Self) }
     }
 }
 
-// TODO: Optimize samplers into parallel operations.
-// This could be deferred into motiongfx::pipeline?
-// See also https://github.com/voxell-tech/motiongfx/issues/72
-
-/// # Panics
-///
-/// Panics if the [`Timeline`] component is sampling itself.
-fn sample_timelines(world: &mut World) {
-    world.try_resource_scope::<MotionGfxWorld, _>(
-        |world, mut motiongfx| {
-            motiongfx.load_pending_timelines(world);
-            motiongfx.sample_timelines(world);
-        },
-    );
-}
-
-/// A unique Id for a [`Timeline`].
-#[derive(Component, Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct TimelineId(u64);
-
-/// Signal for complete timelines
-#[derive(Component)]
-pub struct TimelineComplete;
-
-#[allow(clippy::type_complexity)]
-fn complete_timelines<T>(
-    mut commands: Commands,
-    motiongfx: Res<MotionGfxWorld>,
-    timelines: Query<
-        (Entity, &TimelineId),
-        (With<T>, Without<TimelineComplete>),
-    >,
-) where
-    T: Component,
+impl<S: Component<Mutability = Mutable>> SubjectSource<Entity, S>
+    for BevyWorld
 {
-    for (entity, timeline) in timelines.iter() {
-        if motiongfx
-            .get_timeline(timeline)
-            .is_some_and(|t| t.is_complete())
-        {
-            commands.entity(entity).insert(TimelineComplete);
-        }
+    fn get_source(&self, id: Entity) -> Option<&S> {
+        self.0.get::<S>(id)
+    }
+
+    fn apply_source<R>(
+        &mut self,
+        id: Entity,
+        f: impl FnOnce(&mut S) -> R,
+    ) -> Option<R> {
+        self.0.get_mut::<S>(id).map(|mut m| f(m.as_mut()))
     }
 }
 
-/// Resources that the [`motiongfx`] framework operates on.
-#[derive(Resource)]
-pub struct MotionGfxWorld {
-    id: TimelineId,
-    pending_timelines: HashMap<TimelineId, MutDetect<BevyTimeline>>,
-    timelines: HashMap<TimelineId, MutDetect<BevyTimeline>>,
-    registry: Registry,
-}
-
-impl Default for MotionGfxWorld {
-    fn default() -> Self {
-        Self {
-            id: TimelineId(0),
-            pending_timelines: Default::default(),
-            timelines: Default::default(),
-            registry: Default::default(),
-        }
-    }
-}
-
-impl MotionGfxWorld {
-    pub fn create_builder<W: 'static>(
-        &mut self,
-    ) -> TimelineBuilder<'_, W> {
-        TimelineBuilder::new(&mut self.registry)
-    }
-
-    pub fn create_bevy_builder(&mut self) -> BevyTimelineBuilder<'_> {
-        TimelineBuilder::new(&mut self.registry)
-    }
-
-    pub fn add_timeline(
-        &mut self,
-        timeline: BevyTimeline,
-    ) -> TimelineId {
-        let id = self.id;
-        self.pending_timelines.insert(id, MutDetect::new(timeline));
-
-        self.id.0 = self.id.0.wrapping_add(1);
-        id
-    }
-
-    pub fn remove_timeline(
-        &mut self,
-        id: &TimelineId,
-    ) -> Option<BevyTimeline> {
-        self.timelines
-            .remove(id)
-            .or_else(|| self.pending_timelines.remove(id))
-            .map(|t| t.take())
-    }
-
-    pub fn get_timeline(
+#[cfg(feature = "asset")]
+impl<S: bevy_asset::Asset>
+    SubjectSource<bevy_asset::UntypedAssetId, S> for BevyWorld
+{
+    fn get_source(
         &self,
-        id: &TimelineId,
-    ) -> Option<&BevyTimeline> {
-        self.timelines
-            .get(id)
-            .or_else(|| self.pending_timelines.get(id))
-            .map(|t| &**t)
+        id: bevy_asset::UntypedAssetId,
+    ) -> Option<&S> {
+        self.0
+            .get_resource::<bevy_asset::Assets<S>>()?
+            .get(id.typed::<S>())
     }
 
-    pub fn get_timeline_mut(
+    fn apply_source<R>(
         &mut self,
-        id: &TimelineId,
-    ) -> Option<&mut MutDetect<BevyTimeline>> {
-        self.timelines
-            .get_mut(id)
-            .or_else(|| self.pending_timelines.get_mut(id))
-    }
-
-    pub fn load_pending_timelines(&mut self, world: &mut World) {
-        for (id, mut timeline) in self.pending_timelines.drain() {
-            timeline.bake_actions(
-                &self.registry,
-                BevyWorld::from_ref(world),
-            );
-            self.timelines.insert(id, timeline);
-        }
-    }
-
-    pub fn sample_timelines(&mut self, world: &mut World) {
-        for timeline in
-            self.timelines.values_mut().filter(|t| t.mutated())
-        {
-            timeline.queue_actions();
-            timeline.sample_queued_actions(
-                &self.registry,
-                BevyWorld::from_mut(world),
-            );
-            timeline.reset();
-        }
+        id: bevy_asset::UntypedAssetId,
+        f: impl FnOnce(&mut S) -> R,
+    ) -> Option<R> {
+        self.0
+            .get_resource_mut::<bevy_asset::Assets<S>>()?
+            .into_inner()
+            .get_mut(id.typed::<S>())
+            .map(f)
     }
 }
 
-pub struct MutDetect<T> {
-    inner: T,
-    mutated: bool,
-}
-
-impl<T> MutDetect<T> {
-    pub fn new(inner: T) -> Self {
-        Self {
-            inner,
-            mutated: false,
-        }
-    }
-
-    pub fn mutated(&self) -> bool {
-        self.mutated
-    }
-
-    /// Reset mutation detection flag to `false`.
-    pub fn reset(&mut self) {
-        self.mutated = false
-    }
-
-    pub fn take(self) -> T {
-        self.inner
-    }
-}
-
-impl<T> Deref for MutDetect<T> {
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        &self.inner
-    }
-}
-
-impl<T> DerefMut for MutDetect<T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.mutated = true;
-        &mut self.inner
-    }
-}
+pub type BevyTimeline = Timeline<BevyWorld>;
+pub type BevyTimelineBuilder<'a> = TimelineBuilder<'a, BevyWorld>;
